@@ -2,12 +2,6 @@
   "name": "smart-contracts-admin",
   "version": "0.1.0",
   "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
   "dependencies": {
     "@metaplex-foundation/js": "0.15.0",
     "@metaplex-foundation/mpl-token-metadata": "1.2.5",
@@ -15,30 +9,62 @@
     "@solana/spl-token": "0.1.8",
     "@solana/spl-token-v2": "npm:@solana/spl-token@0.2.0",
     "@solana/wallet-adapter-base": "^0.9.18",
-    "@solana/wallet-adapter-material-ui": "^0.16.20",
-    "@solana/wallet-adapter-phantom": "^0.9.17",
     "@solana/wallet-adapter-react": "^0.15.24",
     "@solana/wallet-adapter-react-ui": "0.9.21",
-    "@solana/wallet-adapter-solflare": "^0.6.18",
-    "@solana/wallet-adapter-torus": "^0.11.23",
     "@solana/wallet-adapter-wallets": "^0.19.5",
     "@solana/web3.js": "^1.66.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "@types/file-saver": "^2.0.5",
-    "@types/node": "18.11.9",
-    "@types/react": "18.0.25",
-    "@types/react-dom": "18.0.8",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.3",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "assert": "^2.0.0",
     "bs58": "^5.0.0",
-    "eslint": "8.27.0",
-    "eslint-config-next": "13.0.3",
+    "crypto-browserify": "^3.12.0",
     "file-saver": "^2.0.5",
-    "next": "13.0.3",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-toastify": "^9.1.1",
-    "typescript": "4.8.4"
+    "react": "^18.2.0",
+    "react-app-rewired": "^2.2.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6.3.0",
+    "react-scripts": "5.0.1",
+    "react-toastify": "9.1.1",
+    "stream-browserify": "^3.0.0",
+    "typescript": "^4.9.3",
+    "util": "^0.12.5",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
+    "eject": "react-app-rewired eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      "chrome >= 67",
+      "edge >= 79",
+      "firefox >= 68",
+      "opera >= 54",
+      "safari >= 14"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "devDependencies": {
     "autoprefixer": "^10.4.13",
+    "eslint": "^8.27.0",
     "postcss": "^8.4.19",
     "tailwindcss": "^3.2.4"
   }
