@@ -31,7 +31,7 @@ export const SOLANA_RPC_HOST_MAINNET = 'https://quiet-aged-frog.solana-mainnet.q
 export const SOLANA_RPC_HOST_DEVNET = 'https://api.devnet.solana.com';
 export const IS_DEV_ENVIRONMENT = false;
 
-export const CONNECTION_NETWORK = IS_DEV_ENVIRONMENT ?  WalletAdapterNetwork.Devnet : WalletAdapterNetwork.Mainnet;
+export const CONNECTION_NETWORK = IS_DEV_ENVIRONMENT ? WalletAdapterNetwork.Devnet : WalletAdapterNetwork.Mainnet;
 export const CONNECTION_NETWORK_RPC = CONNECTION_NETWORK === WalletAdapterNetwork.Mainnet ? SOLANA_RPC_HOST_MAINNET : SOLANA_RPC_HOST_DEVNET;
 
 export const NFT_VAULT_POOL_SEED = "nft_vault_pool";
@@ -46,18 +46,17 @@ export const REACT_APP_SKT_TOKEN_ADDRESS = "SKTsW8KvzopQPdamXsPhvkPfwzTenegv3c3P
 export const REACT_APP_SOLANA_RPC_HOST_MAINNET = "https://quiet-aged-frog.solana-mainnet.quiknode.pro/6a56c0f12de472ff85a245955e5ff33d99704b1a/";
 export const VAULT_ADDRESS = "8Cd1qkkbq9XeB7GzUwVmak9d33QNd3ovqoTBEz8vmYHP";
 
-export const SPLTOKENS_MAP_GET_TOKEN_NAME: (tokenAddress: string) => { tokenAddress: string; tokenName: string } = (tokenAddress: string) =>
-{
-    // @ts-ignore
-    const res = [...SPLTOKENS_MAP].find(([_tokenName, _tokenAddress]) => _tokenAddress == tokenAddress);
+export const SPLTOKENS_MAP_GET_TOKEN_NAME: (tokenAddress: string) => { tokenAddress: string; tokenName: string } = (tokenAddress: string) => {
+  // @ts-ignore
+  const res = [...SPLTOKENS_MAP].find(([_tokenName, _tokenAddress]) => _tokenAddress === tokenAddress);
 
-    return res ? { tokenName: res[0], tokenAddress: res[1] } : { tokenName: "", tokenAddress: "" }
+  return res ? { tokenName: res[0], tokenAddress: res[1] } : { tokenName: "", tokenAddress: "" }
 }
 
 export const DEBUG_API_CALLS = true;
 export const REACT_APP_IS_ACTIVE_SK_ENDPOINT = IS_DEV_ENVIRONMENT // false means use Servica.io (production)
 export const REACT_APP_RAFFLE_AUTH = "tok42412354692119";
 export const REACT_APP_SERVICA_API_ENDPOINT = "https://api.servica.io/extorio/apis/";
-export const REACT_APP_SK_API_ENDPOINT="https://api.solkitties.net/kitties/hooks/";
+export const REACT_APP_SK_API_ENDPOINT = "https://api.solkitties.net/kitties/hooks/";
 
 
